--- src/client/datascience/jupyter/invalidNotebookFileError.ts
@@ -1,17 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-'use strict';
-import { BaseError } from '../../common/errors/types';
-import '../../common/extensions';
-import * as localize from '../../common/utils/localize';
-
-export class InvalidNotebookFileError extends BaseError {
-    constructor(file?: string) {
-        super(
-            'unknown',
-            file
-                ? localize.DataScience.invalidNotebookFileErrorFormat().format(file)
-                : localize.DataScience.invalidNotebookFileError()
-        );
-    }
-}
