--- src/client/datascience/jupyter/jupyterSelfCertsError.ts
@@ -1,11 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-'use strict';
-import { BaseError } from '../../common/errors/types';
-import '../../common/extensions';
-
-export class JupyterSelfCertsError extends BaseError {
-    constructor(message: string) {
-        super('jupyterselfcert', message);
-    }
-}
