+++ src/client/datascience/jupyter/pythonVariableRequester.ts
@@ -10,7 +10,7 @@ import { IDisposable } from '../../common/types';
 import * as localize from '../../common/utils/localize';
 import { DataFrameLoading, GetVariableInfo } from '../constants';
 import { IJupyterVariable, IKernelVariableRequester } from '../types';
-import { JupyterDataRateLimitError } from './jupyterDataRateLimitError';
+import { JupyterDataRateLimitError } from '../errors/jupyterDataRateLimitError';
 import { executeSilently } from './kernels/kernel';
 import { IKernel } from './kernels/types';
